+++ index.js
@@ -1,4 +1,4 @@
 export * from './multi_hyop/index.js'
 export * from './verify_multi_hyop/index.js'
-export * from './single_hyop/index.js'
-export * from './verify_single_hyop/index.js'
+export * from './hyop/index.js'
+export * from './verify_hyop/index.js'