+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # hyop
 
+## 0.3.0
+
+### Minor Changes
+
+- minor:
+
+      + hyop: aliased by single_hyop,singleHyop
+      + verify_hyop: aliased by verify_single_hyop,verifySingleHyop
+      verify_hyop,verify_multi_hyop: unused hyop: console.warn instead of throw Error: fix: conditionally rendered elements which have a hyop
+
+  size-limit:
+
+      verify_hyop .: + 2 B
+      verifySingleHyop .: + 2 B
+      verify_hyop ./verify_hyop: + 2 B
+      verify_multi_hyop .: + 5 B
+      verifyMultiHyop .: + 5 B
+      verify_multi_hyop ./verify_multi_hyop: + 5 B
+
 ## 0.2.6
 
 ### Patch Changes